Changed license to Apache and added boilerplate.
Because this project is built off of a template, there's a lot of legal gray area, and the sklearn-template documentation provides no answers.
First of all, the project came with a BSD 3-Clause License that listed 2 people as the maintainers. There have also been a lot of places that one or both of these people are listed as the copyright owners. I've changed the license to the Apache-2.0 license, and changed any explicit references to the names of the former copyright owners to my own name. (I might've missed some references, but I don't think I did.) I don't know how legally-binding this is, but given that scikit-learn-contrib put this project out there as a TEMPLATE, it seems reasonable that I can modify it however I want. (I give credit to scikit-learn-contrib in the project README.)
Second, the Apache license recommends that I put some boilerplate at the top of my project files. I don't know whether I can/should put this boilerplate at the top of files I didn't write, so for now I have only put it at the top of files that I wrote from scratch.

The choice to use the Apache license was mostly arbitrary.