Added matplotlib cleanup decorator to unit test.
There was a problem: we were creating pyplot figures in a unit test without showing or closing them. This mutates the global pyplot state and could result in multiple figures being displayed in a later test if plt.show() was called. It appears that the decorator, matplotlib.testing.decorators.cleanup, is used to avoid this situation.

I read the source code for cleanup to figure out how to use it. I am not 100% sure that it will work for this project, because it might be for internal use only (i.e. it may rely on matplotlib's testing infrastructure). So that remains to be seen.