Changed how I'm setting the random seed.
Basically what I'm doing here is setting a new precedent: this is how we'll be setting random seeds from now on. This helps to ensure that random seeds for different functions are different.
I will do a writeup of what the actual code style is and why it's a good idea.